+++ packages/r2wc/.eslintrc.json
@@ -2,7 +2,7 @@
   "$schema": "https://json.schemastore.org/eslintrc",
   "root": true,
   "extends": "plugin:@workleap/react-library",
-  "ignorePatterns": ["node_modules/", "dist/", "storybook-static/"],
+  "ignorePatterns": ["node_modules/", "dist/"],
   "rules": {
     "no-trailing-spaces": "error",
     "no-multiple-empty-lines": ["error", { "max": 1, "maxEOF": 0 }]
