+++ packages/r2wc/package.json
@@ -23,8 +23,6 @@
     ],
     "scripts": {
         "build": "tsup --config ./tsup.build.ts",
-        "storybook": "storybook dev -p 6006",
-        "build-storybook": "storybook build",
         "changeset": "changeset add && changeset version",
         "changeset:add": "changeset add",
         "lint": "pnpm run \"/^lint:.*/\"",
@@ -41,28 +39,13 @@
         "@swc/helpers": "0.5.13",
         "@types/react": "18.3.12",
         "@types/react-dom": "18.3.1",
-        "@workleap/browserslist-config": "2.0.1",
         "@workleap/eslint-plugin": "3.2.2",
         "@workleap/swc-configs": "2.2.3",
         "@workleap/tsup-configs": "3.0.6",
         "@workleap/typescript-configs": "3.0.2",
-        "browserslist": "4.24.2",
-        "prettier": "3.3.3",
         "tsconfig-paths-webpack-plugin": "4.1.0",
         "tsup": "8.3.5",
         "typescript": "5.6.3"
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
     },
     "packageManager": "pnpm@9.5.0+sha512.140036830124618d624a2187b50d04289d5a087f326c9edfc0ccd733d76c4f52c3a313d4fc148794a2a9d81553016004e6742e8cf850670268a7387fc220c903"
 }
