+++ packages/r2wc/src/helpers/react/index.ts
@@ -1,17 +1,1 @@
-import { createElement, type HTMLAttributes, useEffect, useRef } from "react";
-import type { WebComponentHTMLElement } from "../../WebComponentHTMLElement.tsx";
-
-export function createWebComponent<Props = unknown, CustomAttributes = unknown>(tagName: keyof JSX.IntrinsicElements) {
-    return function WebComponent(props: { data?: Props } & HTMLAttributes<HTMLElement> & CustomAttributes) {
-        const { data, ...rest } = props;
-        const ref = useRef<WebComponentHTMLElement<Props>>();
-
-        useEffect(() => {
-            if (data !== undefined && ref.current) {
-                ref.current.data = data;
-            }
-        }, [data]);
-
-        return createElement(tagName, { ref, ...rest });
-    };
-}
+export * from "./create-web-component.ts";
